# backend/openapi.yaml
openapi: 3.0.3
info:
  title: WEBTE2 API
  version: 1.0.0
  description: REST API for the WEBTE2 university physics simulation platform.
servers:
  - url: http://localhost:8080/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Human readable message
...
paths:
...
  /session/create:
    post:
      summary: Create a new Octave session
      operationId: createSession
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                    example: 550e8400-e29b-41d4-a716-446655440000
        '401':
          description: Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Bridge error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /session/{id}:
    delete:
      summary: Destroy an Octave session
      operationId: destroySession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session destroyed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          description: Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session not found or already expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...